make migration ready
